Implemented CLI and improved logging
A command-line interface has been added, which is still in development.
The heart of the CLI is in src/cli.ts while the new argument parser is in src/args.ts.
Additionally, logging now logs relative paths instead of absolute paths, and a new LogLevel.Fatal has been added for some CLI conditions.